Migration to JakartaEE (INCOMPLETE!)
Tests run: 778, Failures: 2, Errors: 28, Skipped: 2

+++ ui/src/test/java/de/ipb_halle/lbac/container/service/ContainerNestingServiceTest.java
@@ -24,7 +24,6 @@ import de.ipb_halle.lbac.container.Container;
 import de.ipb_halle.lbac.container.ContainerType;
 import de.ipb_halle.lbac.items.ItemDeployment;
 import de.ipb_halle.testcontainers.PostgresqlContainerExtension;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 import jakarta.inject.Inject;
@@ -144,7 +143,7 @@ public class ContainerNestingServiceTest extends TestBase {
     }
     @SuppressWarnings("unchecked")
     private int getNestedEntries() {
-        List<BigInteger> amountList = (List) entityManagerService.doSqlQuery("SELECT COUNT(*) FROM nested_containers");
+        List<Long> amountList = (List) entityManagerService.doSqlQuery("SELECT COUNT(*) FROM nested_containers");
         return amountList.get(0).intValue();
     }
     @SuppressWarnings("unchecked")
